pyproject.toml
[tool.poetry]
name = "pybitblock"
version = "4.0"
description = "ℙ𝕪𝔹𝕃𝕆ℂ𝕂 𝕚𝕥𝕤 𝕒 𝔹𝕚𝕥𝕔𝕠𝕚𝕟 𝔻𝕒𝕤𝕙𝕓𝕠𝕒𝕣𝕕 𝕨𝕚𝕥𝕙 ℂ𝕪𝕡𝕙𝕖𝕣𝕡𝕦𝕟𝕜 𝕒𝕖𝕤𝕥𝕙𝕖𝕥𝕚𝕔."
license="MIT"
authors = ["curly60e <curly60e@piserver.xyz>", "SN"]
readme="README.md"
repository="https://github.com/curly60e/pyblock"
homepage="https://github.com/curly60e/pyblock"

[tool.poetry.dependencies]
python = "^3.12"
art = "^5.3"
qrcode = "^7.3"
psutil = "^5.8"
simplejson = "^3.17"
certifi = "^2024.7"
chardet = "^4.0"
idna = "^3.7"
python-gnupg = "^0.4.8"
sseclient-py = "^1.7"
urllib3 = "^1.26"
xmltodict = "^0.12"
python-cfonts = "^1.5"
termcolor = "^1.1"
pycoingecko = "^2.2"
protobuf = "^3.18"
robohash = "^1.1"
Pillow = "^10.3"
numpy = "^1.23"
googleapis-common-protos = "^1.52"
pdfminer = "^20191125"
typer = "^0.4"
jq = "^1.2"
html2text = "^2020.1"
pdf2text = "^1.0"
pdf2txt = "^0.7"
embit = "^0.6"
requests = "^2.32"
typer-cli = "^0.0.13"
term-image = "^0.7"
rich = "^13.7"
urwid = "^2.6"
matplotlib = "^3.9"
asciimatics = "^1.15"
plotext = "^5.2"
blessings = "^1.7"
bitcoinlib = "^0.6"
vanity-address = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pyblock = "pybitblock.console:main"
